fix: Resolve build warnings and errors
This commit addresses a variety of build issues, primarily related to compiler warnings and errors that were treated as fatal. The fixes follow several patterns:

- **Signed/Unsigned Comparison:** Corrected signed/unsigned comparison warnings by casting integer literals to the appropriate unsigned type (e.g., `size_t`, `GLenum`) or by using unsigned literals (e.g., `0U`).

- **Pointer-Integer Comparison:** Replaced `NULL` with `nullptr` for pointer comparisons to ensure type safety and resolve comparison errors.

- **Iterator Comparison:** Reverted an `SB_DCHECK_GE` macro comparing iterators in `flat_map.h` to a simple `SB_DCHECK` to avoid compilation errors, as the macro couldn't handle iterator types.

- **Unused Variables:** Added the `[[maybe_unused]]` attribute to variables that were unused in certain build configurations to eliminate warnings.

- **Missing Headers:** Included `starboard/common/check_op.h` where it was missing to resolve undeclared identifier errors.

+++ starboard/nplb/maximum_player_configuration_explorer_test.cc
@@ -69,7 +69,7 @@ TEST_P(MaximumPlayerConfigurationExplorerTest, SunnyDay) {
   for (const auto& video_filename : video_test_files) {
     VideoDmpReader dmp_reader(video_filename,
                               VideoDmpReader::kEnableReadOnDemand);
-    SB_DCHECK_GT(dmp_reader.number_of_video_buffers(), 0);
+    SB_DCHECK_GT(dmp_reader.number_of_video_buffers(), static_cast<size_t>(0));
     if (SbMediaCanPlayMimeAndKeySystem(dmp_reader.video_mime_type().c_str(),
                                        key_system)) {
       supported_player_configs.emplace_back(nullptr, video_filename,
